fix luminace and color setting

+++ PTB/Stim/Make_BarTexture.m
@@ -9,7 +9,7 @@ bar_h = sobj.Bar_heigth_pix;
 bar_w = sobj.Bar_width_pix;
 
 %Texture objext
-im_matrix = uint8(ones(bar_h, bar_w)) * sobj.stimlumi;
+im_matrix = ones(bar_h, bar_w) * sobj.stimlumi;
 
 %Make textue
 im_tex = Screen('MakeTexture', sobj.wPtr, im_matrix);
